Update runkit scripts, use badges fetched through lambdas
